feat: Implement scoring abstraction
