+++ commit_message.txt
@@ -1,1 +1,1 @@
-feat: Implement senior review improvements (input validation, stop words, sentence tokenizer, README updates, and new tests)
+feat: Complete TypeScript migration and update tests
